Rename prepended modules to avoid name clashing.
