--- api/app/models/spree/api/payment.rb
@@ -1,12 +1,0 @@
-module Spree
-  module Api
-    module Payment
-      private
-
-      def after_void
-        super
-        Spree::Webhooks::Endpoints::QueueRequests.call(event: 'payment.void', payload: {})
-      end
-    end
-  end
-end
