--- api/app/models/spree/api/shipment.rb
@@ -1,12 +1,0 @@
-module Spree
-  module Api
-    module Shipment
-      private
-
-      def after_ship
-        super
-        Spree::Webhooks::Endpoints::QueueRequests.call(event: 'shipment.ship', payload: {})
-      end
-    end
-  end
-end
