+++ api/app/models/spree/api/webhooks/payment.rb
@@ -1,0 +1,14 @@
+module Spree
+  module Api
+    module Webhooks
+      module Payment
+        private
+
+        def after_void
+          super
+          Spree::Webhooks::Endpoints::QueueRequests.call(event: 'payment.void', payload: {})
+        end
+      end
+    end
+  end
+end
