+++ api/app/models/spree/api/webhooks/shipment.rb
@@ -1,0 +1,14 @@
+module Spree
+  module Api
+    module Webhooks
+      module Shipment
+        private
+
+        def after_ship
+          super
+          Spree::Webhooks::Endpoints::QueueRequests.call(event: 'shipment.ship', payload: {})
+        end
+      end
+    end
+  end
+end
